backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
aiosmtplib==2.0.2
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
attrs==24.2.0
black==24.10.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
Faker==30.8.0
fastapi==0.115.0
frozenlist==1.5.0
h11==0.14.0
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.0
multidict==6.1.0
mypy-extensions==1.0.0
nanoid==2.0.0
packaging==24.1
pathspec==0.12.1
phonenumbers==8.13.47
platformdirs==4.3.6
propcache==0.2.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
starlette==0.40.0
tomli==2.0.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.0
yarl==1.16.0
